src/main/java/org/example/Instructor.java
package org.example;

import java.util.Arrays;

public class Instructor extends Person implements Teacher{

    public Instructor(long id, String name) {
        super(id, name);
    }

    @Override
    public void teach(Learner learner, double numberOfHours) {
        learner.learn(numberOfHours);
    }

    @Override
    public void lecture(Learner[] learners, double numberOfHours) {
        double numberOfHoursPerLearner = numberOfHours / learners.length;
        Arrays.stream(learners)
                .forEach(learner -> learner.learn(numberOfHoursPerLearner));
    }



}
